requirements.txt
# Core Dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0,<1.26.0  # Constrained for SciPy compatibility
scipy>=1.10.0  # Explicitly add SciPy with compatible version
pillow>=9.5.0
scikit-learn>=1.3.0

# MLflow
mlflow>=2.8.0

# Weights & Biases
wandb>=0.15.0

# Data Loading
kaggle>=1.5.12
kagglehub>=0.2.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Configuration
pyyaml>=6.0

# Utilities
pandas>=2.0.0
tqdm>=4.65.0

# Model Summary (optional)
torchsummary>=1.5.1

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

